Add migration for changing db readability
This renames the server_description column to server_welcome. In addition,
it does the handy thing of changing all datetime objects to strings,
because MySQL can't handle storing time properly.

MySQL will return different datetime values if the system time changes. A
unix epoch is constant, and will not return different values if the system
time changes. A datetime will.