Add create server functionality
This adds a basic route for adding servers. No errors are brought back to
the user, which should happen. Additionally, the end result isn't
particularly spectacular yet. This should be improved later.

This also fixes some dumb bugs in the view that were causing server_key to
be used in 4 different inputs.